symbol table turn into V's

+++ common_test/13_function_def_and_call.v
@@ -2,7 +2,7 @@
 
 fn average_8b(x i8, y i8) i8 {
 	n := 2
-	return (x + y) / n
+	return i8((x + y) / n)
 }
 
 a, b, c := i8(4), i8(7), i8(0)
